Code review and refactoring from acl-enhancements (PR 43) (#44)
This ACL Enhancement is a big change in how the ACL are created and calculated. In this commit, ACL calculations are done using the On-behalf-of flow so that actions take on the storage account are done as the user.  This also allowed us to add groups to the ACL calculation instead of just users.  Changes in both the front and back ends were done to accomodate the group changes.

* Code review comment
* Delete unused imports
* Minor refactorings attempting to reduce the number of operations in foreach loops
Adding lots of comments
* Temporary adding of source of container permissions
* Remove 1 of 2 annotations
* Removed annotation 2 of 2 after validation
* Minor refactoring
* Code comments on TopLevelFolders.cs
* Create toast message for errors
* Improve handling of prereqs in TopLevelFoldersPOST
* Code formatting
* Minor refactoring
* Parallelization of translating UPN/group name to object ID
Minor code formatting
* Correct usage of async parallelization
* Remove mistakenly initialized cToken